+++ Task04/Program.cs
@@ -27,19 +27,18 @@ namespace Task04
         public static void RunTask04()
         {
             int N = int.Parse(Console.ReadLine());
-
-            int A, D;
+            double A, D;
             // TODO: считайте необходимые значения A и D
-            A = int.Parse(Console.ReadLine());
-            D = int.Parse(Console.ReadLine());
+            A = double.Parse(Console.ReadLine());
+            D = double.Parse(Console.ReadLine());
             // TODO: инициализируйте массив и передайте его в соответствующий метод вместе с необходимыми значениями 
-            int[] arr = new int[N];
+            double[] arr = new double[N];
             FillArray(arr, A, D);
             // TODO: выведите массив на экран
             Console.WriteLine(string.Join(" ", arr));
         }
         
-        static void FillArray(int[] arr, int a, int d)
+        static void FillArray(double[] arr, double a, double d)
         {
             for (int i = 0; i < arr.Length; i++)
             {
